+++ src/Controllers/WebhooksController.php
@@ -2,10 +2,9 @@
 
 namespace Solspace\FreeformPayments\Controllers;
 
-use \Stripe\Event;
-use Solspace\FreeformPayments\Controllers\BasePaymentsController;
-use Solspace\FreeformPayments\Integrations\PaymentGateways\Stripe;
 use Solspace\FreeformPayments\FreeformPayments;
+use Solspace\FreeformPayments\Integrations\PaymentGateways\Stripe;
+use Stripe\Event;
 use yii\web\HttpException;
 
 //TODO: create abstract controller
@@ -46,32 +45,33 @@ class WebhooksController extends BasePaymentsController
 
         //TODO: implement all notification service call as events?
         //TODO: update payment/subscription status accordingly
+        $errorMessage = FreeformPayments::t('Event is not linked to freeform submission');
         switch ($event->type) {
             case Event::CHARGE_SUCCEEDED:
                 $submissionId = $event->data->object->metadata->submission;
                 if (!$submissionId) {
-                    throw new HttpException(400, FreeformPayments::t('Event is not linked to freeform submission'));
+                    throw new HttpException(400, $errorMessage);
                 }
                 $this->getPaymentsNotificationService()->sendChargeSucceeded($submissionId);
                 break;
             case Event::CHARGE_FAILED:
                 $submissionId = $event->data->object->metadata->submission;
                 if (!$submissionId) {
-                    throw new HttpException(400, FreeformPayments::t('Event is not linked to freeform submission'));
+                    throw new HttpException(400, $errorMessage);
                 }
                 $this->getPaymentsNotificationService()->sendChargeFailed($submissionId);
                 break;
             case Event::CUSTOMER_SUBSCRIPTION_CREATED:
                 $submissionId = $event->data->object->metadata->submission;
                 if (!$submissionId) {
-                    throw new HttpException(400, FreeformPayments::t('Event is not linked to freeform submission'));
+                    throw new HttpException(400, $errorMessage);
                 }
                 $this->getPaymentsNotificationService()->sendSubscriptionCreated($submissionId);
                 break;
             case Event::CUSTOMER_SUBSCRIPTION_DELETED:
                 $submissionId = $event->data->object->metadata->submission;
                 if (!$submissionId) {
-                    throw new HttpException(400, FreeformPayments::t('Event is not linked to freeform submission'));
+                    throw new HttpException(400, $errorMessage);
                 }
                 $this->getPaymentsNotificationService()->sendSubscriptionEnded($submissionId);
                 break;
@@ -80,7 +80,7 @@ class WebhooksController extends BasePaymentsController
                 //TODO: fix to correct path
                 $submissionId = $event->data->object->lines->data[0]->subscription;
                 if (!$submissionId) {
-                    throw new HttpException(400, FreeformPayments::t('Event is not linked to freeform submission'));
+                    throw new HttpException(400, $errorMessage);
                 }
                 $this->getPaymentsNotificationService()->sendSubscriptionPaymentSucceeded($submissionId);
                 break;
@@ -89,7 +89,7 @@ class WebhooksController extends BasePaymentsController
                 //TODO: fix to correct path
                 $submissionId = $event->data->object->lines->data[0]->subscription;
                 if (!$submissionId) {
-                    throw new HttpException(400, FreeformPayments::t('Event is not linked to freeform submission'));
+                    throw new HttpException(400, $errorMessage);
                 }
                 $this->getPaymentsNotificationService()->sendSubscriptionPaymentFailed($submissionId);
                 break;