+++ src/Library/Traits/ModelCacheTrait.php
@@ -26,6 +26,7 @@ trait ModelCacheTrait
      * Saves model to cache
      *
      * @param Model $model
+     *
      * @return void
      */
     protected function cacheSave(Model $model)
@@ -37,11 +38,14 @@ trait ModelCacheTrait
      * Deletes cached model from cache
      *
      * @param integer $id
+     *
      * @return void
      */
     protected function cacheDelete(int $id)
     {
-        unset($this->modelCache[$model->id]);
+        if (isset($this->modelCache[$id])) {
+            unset($this->modelCache[$id]);
+        }
     }
 
     /**
@@ -49,18 +53,20 @@ trait ModelCacheTrait
      *
      * @return void
      */
-    protected function cacheClear() {
-        $this->modelCache = array();
+    protected function cacheClear()
+    {
+        $this->modelCache = [];
     }
 
     /**
      * Returns cached model or null if id not cached
      *
      * @param integer $id
+     *
      * @return Model|null
      */
     protected function cacheGet(int $id)
     {
-        return isset($this->modelCache[$id]) ? $this->modelCache[$id] : null;
+        return $this->modelCache[$id] ?? null;
     }
 }