+++ src/Resources/Bundles/PaymentsBundle.php
@@ -10,23 +10,6 @@ class PaymentsBundle extends AbstractFreeformPaymentsAssetBundle
     /**
      * @inheritDoc
      */
-    protected function getSourcePath(): string
-    {
-        return '@Solspace/FreeformPayments/Resources';
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getScripts(): array
-    {
-        return [
-        ];
-    }
-
-    /**
-     * @inheritDoc
-     */
     public function getStylesheets(): array
     {
         return ['css/main.css'];