+++ src/translations/en-US/freeform-payments.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Freeform for Craft
+ *
+ * @package       Solspace:Freeform
+ * @author        Solspace, Inc.
+ * @copyright     Copyright (c) 2008-2017, Solspace, Inc.
+ * @link          https://solspace.com/craft/freeform
+ * @license       https://solspace.com/software/license-agreement
+ */
+
+return [
+    'Subscription not found'                     => 'Subscription not found',
+    'Error during subscription cancellation'     => 'Error during subscription cancellation',
+    'Unsubscribed successfully'                  => 'Unsubscribed successfully',
+    'Invalid integration'                        => 'Invalid integration',
+    'Integration is not configured properly'     => 'Integration is not configured properly',
+    'Invalid payload'                            => 'Invalid payload',
+    'Invalid signature'                          => 'Invalid signature',
+    'Event is not linked to freeform submission' => 'Event is not linked to freeform submission',
+    'Fix Missing Payments'                       => 'Fix Missing Payments',
+    'Form does not contain payment fields'       => 'Form does not contain payment fields',
+    'Payments are not set up for the form'       => 'Payments are not set up for the form',
+    'No submissions found'                       => 'No submissions found',
+
+    'Payment gateway is not configured!'   => 'Payment gateway is not configured!',
+    'Payment type is not configured!'      => 'Payment type is not configured!',
+    'Payment amount is not configured!'    => 'Payment amount is not configured!',
+    'Subscription plan is not configured!' => 'Subscription plan is not configured!',
+];