+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Solspace Freeform Payments Changelog
 
+## 1.0.0-beta.2 - 2018-09-14
+### Fixed
+- Fixed a bug where Payments Credit Card field was not showing up correctly in Composer and erroring in front end templates.
+- Fixed a bug where not all currencies were available to use.
+
 ## 1.0.0-beta.1 - 2018-09-13
 ### Added
 - Initial release.